BE: Restructure folders and names, bug fixes
