+++ backend/.env
@@ -2,3 +2,10 @@ PORT=3000
 GITLAB_CLIENT_ID=
 GITLAB_CLIENT_SECRET=
 GITLAB_REDIRECT_URI=http://localhost:3000/auth/gitlab/callback
+OPENAI_API_KEY=
+DB_HOST=
+DB_USER=
+DB_PASSWORD=
+DB_DATABASE=unframe
+DB_PORT=5432
+SESSION_SECRET=
