+++ backend/src/AI/ai-routes.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const aiController = require('../controllers/aiController');
+const aiController = require('./ai-controller');
 const router = express.Router();
 
 router.post('/', aiController.askAI);
