+++ backend/src/authorization/auth-routes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const authController = require('../controllers/authController');
+const authController = require('./auth-controller');
 
 router.get('/gitlab', authController.authenticateGitlab);
 router.get('/gitlab/callback', authController.gitlabCallback);
